move files to correct dir
